[feat[home]]#35 Send detail data 메탈볼 클릭시 id값에 맞는 데이터 전달
bundle로 값 전달, 뷰에 뿌려주기

+++ app/src/main/java/kr/co/nottodo/data/remote/api/home/HomeService.kt
@@ -24,7 +24,7 @@ interface HomeService {
 
     @GET("mission/{missionId}")
     suspend fun getHomeBottomDetail(
-        @Path("missionId") missionId: Int
+        @Path("missionId") missionId: Long
     ): ResponHomeMissionDetail
 
     @GET("mission/week/{startDate}")
